requirements.txt
boto3
prometheus_client
decorator
pyyaml
